'int64' to 'int'

+++ iso.go
@@ -21,7 +21,7 @@ const (
 type iso_date struct {
   
   //
-  iso_year int64
-  iso_week_number int64
-  iso_week_day int64
+  iso_year int
+  iso_week_number int
+  iso_week_day int
 }
